pyproject.toml
[project]
name = "mcpweaver"
version = "0.1.0"
description = "Lightweight YAML configureable MCP Server / Client tools"
readme = "README.md"
authors = [
  {name = "Petrus H Zwart", email = "phzwart@lbl.gov"}
]
maintainers = [
  {name = "Petrus H Zwart", email = "phzwart@lbl.gov"}
]
classifiers = [

]
license = {text = "MIT"}
dependencies = [
  "typer",
  "rich",
  "pyyaml",
  "fastapi",
  "uvicorn[standard]",
  "requests"
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb"    
]

[project.urls]
bugs = "https://github.com/phzwart/mcpweaver/issues"
changelog = "https://github.com/phzwart/mcpweaver/blob/master/changelog.md"
homepage = "https://github.com/phzwart/mcpweaver"

[project.scripts]
mcpweaver = "mcpweaver.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
